nrf serialization: remove "wakeup" command
We did this call through the kernel to get things working, but we don't
strictly need it. The main thing it solves is that it gets us a callback
when we call yield(), and it works on any yield(). To replace it, we
need to do checks anytime we would get to yield and see if we want to
generate the same callback ourselves without going through the kernel.